cnn_assignment/requirements.txt
numpy>=1.21.0
matplotlib>=3.3.0
pillow>=8.0.0
numba>=0.56.0
scipy>=1.7.0
imageio>=2.9.0
scikit-learn>=1.0.0
tqdm>=4.62.0
seaborn>=0.11.0 